# swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "Habitat Commissioner",
    "description": "Express REST API for moving data between orgs and environments in apigee",
    "license": {
      "name": "MIT",
      "url": "http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/proxies/apigee/list": {
      "get": {
        "description": "List all of the proxies in the specified org in Apigee",
        "summary": "List proxies",
        "tags": [
          "api proxies"
        ],
        "operationId": "listProxies",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/apigee/details/{proxy_name}": {
      "get": {
        "description": "Details of a proxy in the specified org in Apigee",
        "summary": "Get proxy",
        "tags": [
          "api proxies"
        ],
        "operationId": "getProxy",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/apigee/list/{proxy_name}": {
      "get": {
        "description": "List the revisions of an proxy in Apigee",
        "summary": "List proxy revisions",
        "tags": [
          "api proxies"
        ],
        "operationId": "listProxyRevisions",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/apigee/details/{proxy_name}/{revision_number}": {
      "get": {
        "description": "Details of a single proxy revision in the specified org in Apigee",
        "summary": "Get proxy revision",
        "tags": [
          "api proxies"
        ],
        "operationId": "getProxyRevision",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          },
          {
            "name": "revision_number",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "revision number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/apigee/deployments/{proxy_name}": {
      "get": {
        "description": "Get each deployment of specified proxy in Apigee",
        "summary": "Get proxy deployments",
        "tags": [
          "api proxies"
        ],
        "operationId": "getProxyDeployments",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/repo/list": {
      "get": {
        "description": "List of all proxy directories in the specified repo",
        "summary": "List proxies in repo",
        "tags": [
          "api proxies"
        ],
        "operationId": "listRepoProxies",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/repo/details/{proxy_name}": {
      "get": {
        "description": "Details of a single proxies configuration file in the repo",
        "summary": "Get a proxy from the repo",
        "tags": [
          "api proxies"
        ],
        "operationId": "getRepoProxy",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/repo": {
      "post": {
        "description": "Export proxies from Apigee into the repo",
        "summary": "Export proxies",
        "tags": [
          "api proxies"
        ],
        "operationId": "exportProxies",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ExportProxiesRequest"
              }
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/apigee": {
      "post": {
        "description": "Import proxies from the repo into Apigee. We also deploy the proxy into the specified environment",
        "summary": "Import proxies",
        "tags": [
          "api proxies"
        ],
        "operationId": "importProxies",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/ApigeeRequest"
            }
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/deploy/{proxy_name}/{revision_number}": {
      "post": {
        "description": "Deploy a revision in Apigee",
        "summary": "Deploy proxy",
        "tags": [
          "api proxies"
        ],
        "operationId": "deployProxy",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "repo",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "repo name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          },
          {
            "name": "revision_number",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "revision number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/undeploy/{proxy_name}/{revision_number}": {
      "delete": {
        "description": "Undeploy a proxy revision",
        "summary": "Undeploy proxy",
        "tags": [
          "api proxies"
        ],
        "operationId": "undeployProxy",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "env",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "env name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          },
          {
            "name": "revision_number",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "revision number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/{proxy_name}/{revision_number}": {
      "delete": {
        "description": "Delete a proxy revision",
        "summary": "Delete proxy revision",
        "tags": [
          "api proxies"
        ],
        "operationId": "deleteProxyRevision",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          },
          {
            "name": "revision_number",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "revision number"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
    "/proxies/{proxy_name}": {
      "delete": {
        "description": "Delete a proxy from Apigee",
        "summary": "Delete proxy",
        "tags": [
          "api proxies"
        ],
        "operationId": "deleteProxy",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "org",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "org name"
          },
          {
            "name": "token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": "access token"
          },
          {
            "name": "proxy_name",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "proxy name"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ApigeeResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ApigeeResponse": {
      "title": "apigeeResponse",
      "type": "object",
      "properties": {
        "code": {
          "description": "",
          "example": 200,
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "description": "",
          "example": "Context sensitive message",
          "type": "string"
        },
        "data": {
          "description": "Generally reflects the response directly from the Apigee management API. May be an object or an array",
          "type": "object"
        }
      }
    },
    "ApigeeRequest": {
      "type": "array",
      "items": {
        "type": "string",
        "description": "item name or ID",
        "example": "item-name-or-id"
      }
    },
...
    "ExportProxiesRequest": {
      "title": "exportProxiesRequest",
      "example": {
        "name": "{{proxy_name}}",
        "revision_number": "{{revision_number}}"
      },
      "type": "object",
      "properties": {
        "name": {
          "description": "",
          "example": "{{proxy_name}}",
          "type": "string"
        },
        "revision_number": {
          "description": "",
          "example": "{{revision_number}}",
          "type": "string"
        }
      },
      "required": [
        "name",
        "revision_number"
      ]
    },
...
  }
}